Updated.
We are now showing the list of patients. Also a few files were added on the front end to add edit patient functionality. Adding a new patient functionality is now on a separate page as well. A delete button was added next to the edit button to delete a patient.